tried to get picture as input
dimag finish
